# Re-implement data generation for trainers