--- a/include/struct_defs/trainer_data.h
+++ b/include/struct_defs/trainer_data.h
@@ -1,11 +1,16 @@
 #ifndef POKEPLATINUM_STRUCT_TRAINER_DATA_H
 #define POKEPLATINUM_STRUCT_TRAINER_DATA_H
 
-#include "constants/string.h"
+/*
+ * Note: *most* source files should not include this header directly, and should
+ * instead including `struct_defs/trainer.h`. This header is held separately for
+ * use by data-packing routines.
+ */
 
-#include "struct_defs/sentence.h"
+#include "constants/moves.h"
 
-#include "strbuf.h"
+#define MAX_TRAINER_ITEMS 4
+#define MAX_DV            255
 
 enum TrainerDataType {
     TRDATATYPE_BASE = 0,
@@ -14,71 +19,46 @@ enum TrainerDataType {
     TRDATATYPE_WITH_MOVES_AND_ITEM,
 };
 
-/**
- * @brief Trainer data encapsulation. Contains all trainer-level data needed to instantiate a trainer battle.
- */
-typedef struct Trainer {
-    u8 type; //< The data type of the trainer; controls which TrainerMon struct is used for the party.
-    u8 class; //< The trainer's class, e.g. "Youngster" or "Gym Leader."
-    u8 sprite; //< Which sprite ID the trainer uses.
-    u8 partySize; //< How many Pokemon the trainer has in their party.
+typedef struct TrainerHeader {
+    u8 monDataType;
+    u8 trainerType;
+    u8 sprite;
+    u8 partySize;
+    u16 items[MAX_TRAINER_ITEMS];
+    u32 aiMask;
+    u32 battleType;
+} TrainerHeader;
 
-    u16 items[4]; //< Items that the trainer has in their bag.
-
-    u32 aiMask; //< Mask of AI flags to use for this trainer in battle.
-
-    u32 battleType; //< The type of battle to be instantiated.
-
-    charcode_t name[TRAINER_NAME_LEN + 1]; //< The trainer's name.
-
-    Sentence winMsg; //< The message to display if the trainer wins.
-    Sentence loseMsg; //< The message to display if the trainer loses.
-} Trainer;
-
-#define MAX_DV 255
-
-/**
- * @brief The most basic Trainer Pokemon structure. No item is held, and moves are populated from the learnset.
- */
 typedef struct TrainerMonBase {
-    u16 dv; //< Difficulty value; determines how many IVs the Pokemon will have in each stat.
-    u16 level; //< The Pokemon's level.
-    u16 species; //< Composite mask of the Pokemon's level (least-significant 10 bits) and form (most-significant 6 bits).
-    u16 cbSeal; //< Ball seal applied to the Pokemon.
+    u16 dv;
+    u16 level;
+    u16 species;
+    u16 cbSeal;
 } TrainerMonBase;
 
-/**
- * @brief Trainer Pokemon that knows a certain set of moves.
- */
 typedef struct TrainerMonWithMoves {
-    u16 dv; //< Difficulty value; determines how many IVs the Pokemon will have in each stat.
-    u16 level; //< The Pokemon's level.
-    u16 species; //< Composite mask of the Pokemon's level (least-significant 10 bits) and form (most-significant 6 bits).
-    u16 moves[4]; //< The Pokemon's known moves.
-    u16 cbSeal; //< Ball seal applied to the Pokemon.
+    u16 dv;
+    u16 level;
+    u16 species;
+    u16 moves[LEARNED_MOVES_MAX];
+    u16 cbSeal;
 } TrainerMonWithMoves;
 
-/**
- * @brief Trainer Pokemon that has a specific held item.
- */
 typedef struct TrainerMonWithItem {
-    u16 dv; //< Difficulty value; determines how many IVs the Pokemon will have in each stat.
-    u16 level; //< The Pokemon's level.
-    u16 species; //< Composite mask of the Pokemon's level (least-significant 10 bits) and form (most-significant 6 bits).
-    u16 item; //< The Pokemon's held item.
-    u16 cbSeal; //< Ball seal applied to the Pokemon.
+    u16 dv;
+    u16 level;
+    u16 species;
+    u16 item;
+    u16 cbSeal;
 } TrainerMonWithItem;
 
-/**
- * @brief Trainer Pokemon that has a specific held item and knows a certain set of moves.
- */
 typedef struct TrainerMonWithMovesAndItem {
-    u16 dv; //< Difficulty value; determines how many IVs the Pokemon will have in each stat.
-    u16 level; //< The Pokemon's level.
-    u16 species; //< Composite mask of the Pokemon's level (least-significant 10 bits) and form (most-significant 6 bits).
-    u16 item; //< The Pokemon's held item.
-    u16 moves[4]; //< The Pokemon's known moves.
-    u16 cbSeal; //< Ball seal applied to the Pokemon.
+    u16 dv;
+    u16 level;
+    u16 species;
+    u16 item;
+    u16 moves[LEARNED_MOVES_MAX];
+    u16 cbSeal;
 } TrainerMonWithMovesAndItem;
 
 #endif // POKEPLATINUM_STRUCT_TRAINER_DATA_H