--- a/include/struct_defs/trainer_data.h
+++ b/include/struct_defs/trainer_data.h
@@ -9,6 +9,8 @@
 
 #include "constants/moves.h"
 
+#define TRAINER_MON_FORM_SHIFT 10
+
 #define MAX_TRAINER_ITEMS 4
 #define MAX_DV            255