--- a/res/field/scripts/scripts_battleground.s
+++ b/res/field/scripts/scripts_battleground.s
@@ -397,55 +397,55 @@ _0605:
     Return
 
 _06B0:
-    SetVar 0x4005, trainer_leader_roark_rematch 
+    SetVar 0x4005, TRAINER_LEADER_ROARK_REMATCH 
     Return
 
 _06B8:
-    SetVar 0x4005, trainer_leader_gardenia_rematch
+    SetVar 0x4005, TRAINER_LEADER_GARDENIA_REMATCH
     Return
 
 _06C0:
-    SetVar 0x4005, trainer_leader_wake_rematch
+    SetVar 0x4005, TRAINER_LEADER_WAKE_REMATCH
     Return
 
 _06C8:
-    SetVar 0x4005, trainer_leader_maylene_rematch
+    SetVar 0x4005, TRAINER_LEADER_MAYLENE_REMATCH
     Return
 
 _06D0:
-    SetVar 0x4005, trainer_leader_fantina_rematch
+    SetVar 0x4005, TRAINER_LEADER_FANTINA_REMATCH
     Return
 
 _06D8:
-    SetVar 0x4005, trainer_leader_candice_rematch
+    SetVar 0x4005, TRAINER_LEADER_CANDICE_REMATCH
     Return
 
 _06E0:
-    SetVar 0x4005, trainer_leader_byron_rematch
+    SetVar 0x4005, TRAINER_LEADER_BYRON_REMATCH
     Return
 
 _06E8:
-    SetVar 0x4005, trainer_leader_volkner_rematch
+    SetVar 0x4005, TRAINER_LEADER_VOLKNER_REMATCH
     Return
 
 _06F0:
-    SetVar 0x4005, trainer_cheryl_battleground
+    SetVar 0x4005, TRAINER_CHERYL_BATTLEGROUND
     Return
 
 _06F8:
-    SetVar 0x4005, trainer_riley_battleground
+    SetVar 0x4005, TRAINER_RILEY_BATTLEGROUND
     Return
 
 _0700:
-    SetVar 0x4005, trainer_marley_battleground
+    SetVar 0x4005, TRAINER_MARLEY_BATTLEGROUND
     Return
 
 _0708:
-    SetVar 0x4005, trainer_buck_battleground
+    SetVar 0x4005, TRAINER_BUCK_BATTLEGROUND
     Return
 
 _0710:
-    SetVar 0x4005, trainer_mira_battleground
+    SetVar 0x4005, TRAINER_MIRA_BATTLEGROUND
     Return
 
 _0718: