--- a/res/field/scripts/scripts_canalave_city_gym.s
+++ b/res/field/scripts/scripts_canalave_city_gym.s
@@ -22,7 +22,7 @@ _0016:
     CreateJournalEvent LOCATION_EVENT_GYM_WAS_TOO_TOUGH, 35, 0, 0, 0
     Message 0
     CloseMessage
-    StartTrainerBattle trainer_leader_byron
+    StartTrainerBattle TRAINER_LEADER_BYRON
     CheckWonBattle 0x800C
     GoToIfEq 0x800C, 0, _0120
     Message 1
@@ -32,14 +32,14 @@ _0016:
     WaitSound
     GiveBadge BADGE_ID_MINE
     ScrCmd_260 23
-    SetTrainerFlag trainer_black_belt_david
-    SetTrainerFlag trainer_worker_jackson
-    SetTrainerFlag trainer_worker_gary
-    SetTrainerFlag trainer_ace_trainer_cesar
-    SetTrainerFlag trainer_ace_trainer_breanna
-    SetTrainerFlag trainer_black_belt_ricky
-    SetTrainerFlag trainer_worker_gerardo
-    CreateJournalEvent LOCATION_EVENT_BEAT_GYM_LEADER, 35, trainer_leader_byron, 0, 0
+    SetTrainerFlag TRAINER_BLACK_BELT_DAVID
+    SetTrainerFlag TRAINER_WORKER_JACKSON
+    SetTrainerFlag TRAINER_WORKER_GARY
+    SetTrainerFlag TRAINER_ACE_TRAINER_CESAR
+    SetTrainerFlag TRAINER_ACE_TRAINER_BREANNA
+    SetTrainerFlag TRAINER_BLACK_BELT_RICKY
+    SetTrainerFlag TRAINER_WORKER_GERARDO
+    CreateJournalEvent LOCATION_EVENT_BEAT_GYM_LEADER, 35, TRAINER_LEADER_BYRON, 0, 0
     ClearFlag 0x1B2
     ClearFlag 0x2CC
     ClearFlag 0x1B8