--- a/res/field/scripts/scripts_eterna_city_gym.s
+++ b/res/field/scripts/scripts_eterna_city_gym.s
@@ -68,7 +68,7 @@ _0094:
     CreateJournalEvent LOCATION_EVENT_GYM_WAS_TOO_TOUGH, 67, 0, 0, 0
     Message 0
     CloseMessage
-    StartTrainerBattle trainer_leader_gardenia
+    StartTrainerBattle TRAINER_LEADER_GARDENIA
     CheckWonBattle 0x800C
     GoToIfEq 0x800C, 0, _01C3
     Message 1
@@ -78,12 +78,12 @@ _0094:
     WaitSound
     GiveBadge BADGE_ID_FOREST
     ScrCmd_260 23
-    SetTrainerFlag trainer_aroma_lady_jenna
-    SetTrainerFlag trainer_aroma_lady_angela
-    SetTrainerFlag trainer_lass_caroline
-    SetTrainerFlag trainer_beauty_lindsay
+    SetTrainerFlag TRAINER_AROMA_LADY_JENNA
+    SetTrainerFlag TRAINER_AROMA_LADY_ANGELA
+    SetTrainerFlag TRAINER_LASS_CAROLINE
+    SetTrainerFlag TRAINER_BEAUTY_LINDSAY
     ClearFlag 0x1FC
-    CreateJournalEvent LOCATION_EVENT_BEAT_GYM_LEADER, 67, trainer_leader_gardenia, 0, 0
+    CreateJournalEvent LOCATION_EVENT_BEAT_GYM_LEADER, 67, TRAINER_LEADER_GARDENIA, 0, 0
     Message 3
     GoTo _0177
     End
@@ -150,10 +150,10 @@ _01C9:
     LockAll
     FacePlayer
     GoToIfGe 0x40AE, 1, _0211
-    ScrCmd_0EA trainer_lass_caroline
+    ScrCmd_0EA TRAINER_LASS_CAROLINE
     Message 6
     CloseMessage
-    StartTrainerBattle trainer_lass_caroline
+    StartTrainerBattle TRAINER_LASS_CAROLINE
     CheckWonBattle 0x800C
     GoToIfEq 0x800C, 0, _01C3
     Message 7
@@ -176,12 +176,12 @@ _021C:
     LockAll
     FacePlayer
     GoToIfGe 0x40AE, 2, _0270
-    ScrCmd_0EA trainer_aroma_lady_jenna
+    ScrCmd_0EA TRAINER_AROMA_LADY_JENNA
     SetVar 0x8007, 3
     Call _02DA
     Message 8
     CloseMessage
-    StartTrainerBattle trainer_aroma_lady_jenna
+    StartTrainerBattle TRAINER_AROMA_LADY_JENNA
     CheckWonBattle 0x800C
     GoToIfEq 0x800C, 0, _01C3
     Message 9
@@ -204,12 +204,12 @@ _027B:
     LockAll
     FacePlayer
     GoToIfGe 0x40AE, 3, _02CF
-    ScrCmd_0EA trainer_aroma_lady_angela
+    ScrCmd_0EA TRAINER_AROMA_LADY_ANGELA
     SetVar 0x8007, 4
     Call _02DA
     Message 10
     CloseMessage
-    StartTrainerBattle trainer_aroma_lady_angela
+    StartTrainerBattle TRAINER_AROMA_LADY_ANGELA
     CheckWonBattle 0x800C
     GoToIfEq 0x800C, 0, _01C3
     Message 11