--- a/res/field/scripts/scripts_hearthome_city_gym_leader_room.s
+++ b/res/field/scripts/scripts_hearthome_city_gym_leader_room.s
@@ -60,7 +60,7 @@ _00BB:
     Message 0
     CloseMessage
     SetFlag 142
-    StartTrainerBattle trainer_leader_fantina
+    StartTrainerBattle TRAINER_LEADER_FANTINA
     ClearFlag 142
     CheckWonBattle 0x800C
     GoToIfEq 0x800C, 0, _0248
@@ -71,15 +71,15 @@ _00BB:
     WaitSound
     GiveBadge BADGE_ID_RELIC
     ScrCmd_260 23
-    SetTrainerFlag trainer_camper_drew
-    SetTrainerFlag trainer_ace_trainer_allen
-    SetTrainerFlag trainer_ace_trainer_catherine
-    SetTrainerFlag trainer_lass_molly
-    SetTrainerFlag trainer_picnicker_cheyenne
-    SetTrainerFlag trainer_school_kid_chance
-    SetTrainerFlag trainer_school_kid_mackenzie
-    SetTrainerFlag trainer_youngster_donny
-    CreateJournalEvent LOCATION_EVENT_BEAT_GYM_LEADER, 91, trainer_leader_fantina, 0, 0
+    SetTrainerFlag TRAINER_CAMPER_DREW
+    SetTrainerFlag TRAINER_ACE_TRAINER_ALLEN
+    SetTrainerFlag TRAINER_ACE_TRAINER_CATHERINE
+    SetTrainerFlag TRAINER_LASS_MOLLY
+    SetTrainerFlag TRAINER_PICNICKER_CHEYENNE
+    SetTrainerFlag TRAINER_SCHOOL_KID_CHANCE
+    SetTrainerFlag TRAINER_SCHOOL_KID_MACKENZIE
+    SetTrainerFlag TRAINER_YOUNGSTER_DONNY
+    CreateJournalEvent LOCATION_EVENT_BEAT_GYM_LEADER, 91, TRAINER_LEADER_FANTINA, 0, 0
     SetVar 0x407B, 1
     SetFlag 0x206
     ClearFlag 0x207