--- a/res/field/scripts/scripts_oreburgh_city_gym.s
+++ b/res/field/scripts/scripts_oreburgh_city_gym.s
@@ -17,7 +17,7 @@ _000E:
     CreateJournalEvent LOCATION_EVENT_GYM_WAS_TOO_TOUGH, 47, 0, 0, 0
     Message 0
     CloseMessage
-    StartTrainerBattle trainer_leader_roark
+    StartTrainerBattle TRAINER_LEADER_ROARK
     CheckWonBattle 0x800C
     GoToIfEq 0x800C, 0, _0119
     Message 1
@@ -25,17 +25,17 @@ _000E:
     Message 2
     PlaySound SEQ_BADGE
     WaitSound
-    SetTrainerFlag trainer_youngster_jonathon
-    SetTrainerFlag trainer_youngster_darius
+    SetTrainerFlag TRAINER_YOUNGSTER_JONATHON
+    SetTrainerFlag TRAINER_YOUNGSTER_DARIUS
     GiveBadge BADGE_ID_COAL
     ScrCmd_260 23
-    SetTrainerFlag trainer_youngster_jonathon
-    SetTrainerFlag trainer_youngster_darius
+    SetTrainerFlag TRAINER_YOUNGSTER_JONATHON
+    SetTrainerFlag TRAINER_YOUNGSTER_DARIUS
     SetFlag 0x23F
     SetVar 0x40F0, 1
     SetVar 0x4076, 1
     SetVar 0x4079, 2
-    CreateJournalEvent LOCATION_EVENT_BEAT_GYM_LEADER, 47, trainer_leader_roark, 0, 0
+    CreateJournalEvent LOCATION_EVENT_BEAT_GYM_LEADER, 47, TRAINER_LEADER_ROARK, 0, 0
     SetVar 0x4077, 3
     ClearFlag 0x17A
     ClearFlag 0x19D