… generated/ai_flags.txt
… generated/trainers.txt
… res/field/scripts/scripts_pokemon_mansion.s
… res/field/scripts/scripts_restaurant.s
… res/field/scripts/scripts_snowpoint_city_gym.s
… res/field/scripts/scripts_stark_mountain_room_1.s
… res/field/scripts/scripts_sunyshore_city_gym_room_3.s
… res/field/scripts/scripts_survival_area.s
… res/field/scripts/scripts_team_galactic_eterna_building_4f.s
… res/field/scripts/scripts_trainers_school.s
… res/field/scripts/scripts_valley_windworks_building.s
… res/field/scripts/scripts_valley_windworks_outside.s
… res/field/scripts/scripts_valor_cavern.s
… res/field/scripts/scripts_valor_lakefront.s
… res/field/scripts/scripts_veilstone_city_gym.s
… res/meson.build
… res/trainers/data/ace_trainer_abel.json
… res/trainers/data/ace_trainer_alicia.json
… res/trainers/data/ace_trainer_allen.json
… res/trainers/data/ace_trainer_alyssa.json
… res/trainers/data/ace_trainer_anton.json
… res/trainers/data/ace_trainer_arthur.json
… res/trainers/data/ace_trainer_blake.json
… res/trainers/data/ace_trainer_brandi.json
… res/trainers/data/ace_trainer_breanna.json
… res/trainers/data/ace_trainer_brenda.json
… res/trainers/data/ace_trainer_brenna.json
… res/trainers/data/ace_trainer_catherine.json
… res/trainers/data/ace_trainer_cesar.json
… res/trainers/data/ace_trainer_clarice.json